Support workspace config based recommendations
