+++ src/vs/workbench/contrib/tags/electron-browser/workspaceTags.ts
@@ -10,127 +10,13 @@ import { IFileService, IFileStat } from 'vs/platform/files/common/files';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
 import { IWorkbenchContribution } from 'vs/workbench/common/contributions';
-import { endsWith } from 'vs/base/common/strings';
 import { ITextFileService, } from 'vs/workbench/services/textfile/common/textfiles';
 import { ISharedProcessService } from 'vs/platform/ipc/electron-browser/sharedProcessService';
 import { IWorkspaceTagsService, Tags } from 'vs/workbench/contrib/tags/common/workspaceTags';
 import { IWorkspaceInformation } from 'vs/platform/diagnostics/common/diagnostics';
 import { IRequestService } from 'vs/platform/request/common/request';
 import { isWindows } from 'vs/base/common/platform';
-
-const SshProtocolMatcher = /^([^@:]+@)?([^:]+):/;
-const SshUrlMatcher = /^([^@:]+@)?([^:]+):(.+)$/;
-const AuthorityMatcher = /^([^@]+@)?([^:]+)(:\d+)?$/;
-const SecondLevelDomainMatcher = /([^@:.]+\.[^@:.]+)(:\d+)?$/;
-const RemoteMatcher = /^\s*url\s*=\s*(.+\S)\s*$/mg;
-const AnyButDot = /[^.]/g;
-const SecondLevelDomainWhitelist = [
-	'github.com',
-	'bitbucket.org',
-	'visualstudio.com',
-	'gitlab.com',
-	'heroku.com',
-	'azurewebsites.net',
-	'ibm.com',
-	'amazon.com',
-	'amazonaws.com',
-	'cloudapp.net',
-	'rhcloud.com',
-	'google.com',
-	'azure.com'
-];
-
-function stripLowLevelDomains(domain: string): string | null {
-	const match = domain.match(SecondLevelDomainMatcher);
-	return match ? match[1] : null;
-}
-
-function extractDomain(url: string): string | null {
-	if (url.indexOf('://') === -1) {
-		const match = url.match(SshProtocolMatcher);
-		if (match) {
-			return stripLowLevelDomains(match[2]);
-		} else {
-			return null;
-		}
-	}
-	try {
-		const uri = URI.parse(url);
-		if (uri.authority) {
-			return stripLowLevelDomains(uri.authority);
-		}
-	} catch (e) {
-		// ignore invalid URIs
-	}
-	return null;
-}
-
-export function getDomainsOfRemotes(text: string, whitelist: string[]): string[] {
-	const domains = new Set<string>();
-	let match: RegExpExecArray | null;
-	while (match = RemoteMatcher.exec(text)) {
-		const domain = extractDomain(match[1]);
-		if (domain) {
-			domains.add(domain);
-		}
-	}
-
-	const whitemap = whitelist.reduce((map, key) => {
-		map[key] = true;
-		return map;
-	}, Object.create(null));
-
-	const elements: string[] = [];
-	domains.forEach(e => elements.push(e));
-
-	return elements
-		.map(key => whitemap[key] ? key : key.replace(AnyButDot, 'a'));
-}
-
-function stripPort(authority: string): string | null {
-	const match = authority.match(AuthorityMatcher);
-	return match ? match[2] : null;
-}
-
-function normalizeRemote(host: string | null, path: string, stripEndingDotGit: boolean): string | null {
-	if (host && path) {
-		if (stripEndingDotGit && endsWith(path, '.git')) {
-			path = path.substr(0, path.length - 4);
-		}
-		return (path.indexOf('/') === 0) ? `${host}${path}` : `${host}/${path}`;
-	}
-	return null;
-}
-
-function extractRemote(url: string, stripEndingDotGit: boolean): string | null {
-	if (url.indexOf('://') === -1) {
-		const match = url.match(SshUrlMatcher);
-		if (match) {
-			return normalizeRemote(match[2], match[3], stripEndingDotGit);
-		}
-	}
-	try {
-		const uri = URI.parse(url);
-		if (uri.authority) {
-			return normalizeRemote(stripPort(uri.authority), uri.path, stripEndingDotGit);
-		}
-	} catch (e) {
-		// ignore invalid URIs
-	}
-	return null;
-}
-
-export function getRemotes(text: string, stripEndingDotGit: boolean = false): string[] {
-	const remotes: string[] = [];
-	let match: RegExpExecArray | null;
-	while (match = RemoteMatcher.exec(text)) {
-		const remote = extractRemote(match[1], stripEndingDotGit);
-		if (remote) {
-			remotes.push(remote);
-		}
-	}
-	return remotes;
-}
+import { getRemotes, SecondLevelDomainWhitelist, getDomainsOfRemotes } from 'vs/platform/extensionManagement/common/configRemotes';
 
 export function getHashedRemotesFromConfig(text: string, stripEndingDotGit: boolean = false): string[] {
 	return getRemotes(text, stripEndingDotGit).map(r => {
